Address dumbbellplot review feedback
