+++ src/cnsplots/_agent_skill/cnsplots/references/plot-catalog.md
@@ -10,6 +10,7 @@ signature and docstring before writing code.
 - `stripplot`: individual observations across categories.
 - `barplot`: group estimates displayed as bars.
 - `lollipopplot`: values or summaries with stems and markers.
+- `dumbbellplot`: two endpoints per category connected by a line.
 - `histplot`: binned univariate or bivariate distributions.
 - `kdeplot`: smoothed density estimates.
 - `distplot`: compact distribution visualization.
@@ -22,7 +23,6 @@ signature and docstring before writing code.
 - `regplot`: fitted relationship with regression context.
 - `lineplot`: trends or repeated measurements over an ordered axis.
 - `slopeplot`: changes between conditions or time points.
-- `dumbbellplot`: two endpoints per category connected by a line.
 
 ## Matrices and classifications
