+++ tests/test_plots_basic.py
@@ -8,6 +8,7 @@ import numpy as np
 import pandas as pd
 import pytest
 from cycler import cycler
+from matplotlib.collections import PathCollection
 from matplotlib.colors import to_hex, to_rgba
 from matplotlib.patches import (
     Circle,
@@ -551,7 +552,7 @@ def test_dumbbellplot_draws_two_points_per_category() -> None:
     )
 
     cns.figure(120, 150)
-    ax = cns.dumbbellplot(data, x="pathway", y="score", hue="condition")
+    ax = cns.dumbbellplot(data, x="score", y="pathway", hue="condition")
     assert ax is plt.gca()
     assert len(ax.lines) == 3
     assert len(ax.collections) == 2
@@ -563,6 +564,8 @@ def test_dumbbellplot_draws_two_points_per_category() -> None:
     legend = ax.get_legend()
     assert legend is not None
     assert [text.get_text() for text in legend.get_texts()] == ["before", "after"]
+    assert ax.get_xlabel() == "score"
+    assert ax.get_ylabel() == "pathway"
 
 
 def test_dumbbellplot_custom_order_hue_order_and_ax() -> None:
@@ -577,8 +580,8 @@ def test_dumbbellplot_custom_order_hue_order_and_ax() -> None:
 
     result = cns.dumbbellplot(
         data,
-        x="pathway",
-        y="score",
+        x="score",
+        y="pathway",
         hue="condition",
         order=["C", "A", "B"],
         hue_order=["after", "before"],
@@ -591,75 +594,22 @@ def test_dumbbellplot_custom_order_hue_order_and_ax() -> None:
     legend = ax.get_legend()
     assert legend is not None
     assert [text.get_text() for text in legend.get_texts()] == ["after", "before"]
-
-
-def test_dumbbellplot_validates_inputs() -> None:
-    data = pd.DataFrame(
-        {
-            "pathway": ["A", "A", "B", "B", "C", "C"],
-            "score": [1.0, 3.0, 2.0, 5.0, 4.0, 6.0],
-            "condition": ["before", "after"] * 3,
-        }
+    np.testing.assert_allclose(
+        [line.get_xdata() for line in ax.lines],
+        [[6.0, 4.0], [3.0, 1.0], [5.0, 2.0]],
     )
-    with pytest.raises(TypeError, match="must be a pandas DataFrame"):
-        cns.dumbbellplot(cast(Any, []), x="pathway", y="score", hue="condition")
-    with pytest.raises(ValueError, match="not found in data"):
-        cns.dumbbellplot(data, x="missing", y="score", hue="condition")
-    with pytest.raises(ValueError, match="Data is empty"):
-        cns.dumbbellplot(data.iloc[0:0], x="pathway", y="score", hue="condition")
-    with pytest.raises(ValueError, match="must be numeric"):
-        cns.dumbbellplot(
-            data.assign(score=data["condition"]),
-            x="pathway",
-            y="score",
-            hue="condition",
-        )
-    with pytest.raises(ValueError, match="Null values found"):
-        cns.dumbbellplot(
-            data.assign(score=[1.0, None, 2.0, 3.0, 4.0, 5.0]),
-            x="pathway",
-            y="score",
-            hue="condition",
-        )
-    with pytest.raises(ValueError, match="exactly 2 unique values"):
-        cns.dumbbellplot(
-            data[data["condition"] == "before"],
-            x="pathway",
-            y="score",
-            hue="condition",
-        )
-    with pytest.raises(ValueError, match="hue_order"):
-        cns.dumbbellplot(
-            data,
-            x="pathway",
-            y="score",
-            hue="condition",
-            hue_order=["before", "before"],
-        )
-    with pytest.raises(ValueError, match="not present in data"):
-        cns.dumbbellplot(
-            data,
-            x="pathway",
-            y="score",
-            hue="condition",
-            order=["A", "D"],
-        )
-    duplicate = pd.concat([data, data.iloc[[0]]], ignore_index=True)
-    with pytest.raises(ValueError, match="exactly one"):
-        cns.dumbbellplot(
-            duplicate,
-            x="pathway",
-            y="score",
-            hue="condition",
-        )
-    incomplete = data.drop(index=[3])
-    with pytest.raises(ValueError, match="exactly one"):
-        cns.dumbbellplot(
-            incomplete,
-            x="pathway",
-            y="score",
-            hue="condition",
-        )
+    np.testing.assert_allclose(
+        np.asarray(ax.collections[0].get_offsets(), dtype=float),
+        [[6.0, 0.0], [3.0, 1.0], [5.0, 2.0]],
+    )
+    np.testing.assert_allclose(
+        np.asarray(ax.collections[1].get_offsets(), dtype=float),
+        [[4.0, 0.0], [1.0, 1.0], [2.0, 2.0]],
+    )
+    assert ax.lines[0].get_linewidth() == 2
+    for collection in ax.collections:
+        assert isinstance(collection, PathCollection)
+        np.testing.assert_allclose(collection.get_sizes(), [12.0])
 
 
 def test_placeholderplot_renders_centered_placeholder() -> None:
