+++ tests/typing_contract.py
@@ -39,7 +39,7 @@ if TYPE_CHECKING:
         assert_type(cns.lineplot(data, x="x", y="y", ax=ax), Axes)
         assert_type(cns.regplot(data, x="x", y="y", color=(0.1, 0.2, 0.3)), Axes)
         assert_type(
-            cns.dumbbellplot(data, x="group", y="value", hue="condition", ax=ax),
+            cns.dumbbellplot(data, x="value", y="group", hue="condition", ax=ax),
             Axes,
         )
